Fix exporting and importing data
Do note that there is a current error in some Android versions where it can't read files from certain locations, namely from Download, Picture, Video, Audio and Recent. Which is on the "Open From" side dialogue, this will make the file path cut short, not including the CSV, which will error out.

It is recommended to choose the expenses file from the device's internal storage or the SD card if installed, as it will have the full path.